UI : Code cleanup (relative include paths), warning fixes.
git-svn-id: file:///home/michael/svn/yuni/trunk@2450 7e728022-c0ab-4603-8e65-c666ead28a33

+++ src/yuni/ui/gl/view.h
@@ -114,7 +114,7 @@ namespace UI
 		EventPropagation doKeyUp(Input::Key key);
 
 		//! Get the top-most control at given coordinates. (Mainly useful for clicking)
-		IControl* getControlAt(float x, float y);
+		IControl* getControlAt(int x, int y);
 
 		/*!
 		** \brief Draw the view
